moved files to src>components>Ingress
