+++ src/components/PaymentScreen/CreditCardForm.js
@@ -1,7 +1,7 @@
 import Card from "@comeon/react-credit-card";
 import React from "react";
 import styled from "styled-components";
-import Input from "../../../components/Form/Input";
+import Input from "../Form/Input";
 
 export default function CreditCardForm() {
   const [values, setValues] = React.useState({
