--- src/pages/Dashboard/Payment/ReviewPayment.js
@@ -1,5 +1,0 @@
-export default function ReviewPayment() {
-  return (
-    <>Você já pagou</>
-  );
-};
